+++ src/media/color/rgb/impls/u8.rs
@@ -1,8 +1,9 @@
 // devela::media::color::rgb::impls::u8
 
-use crate::{Norm, Rgb8, Rgb16, RgbF32, RgbF64, Rgba8, Rgba16, RgbaF32, RgbaF64};
+use crate::{Norm, Rgb8, Rgb16, RgbF32, RgbF64, Rgba8, RgbaF32, RgbaF64};
 #[cfg(doc)]
 use crate::{Rgb, Rgba};
+// use crate::{Rgba16};
 
 #[allow(missing_docs)]
 #[rustfmt::skip]